# contracts/phase11.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Communications Service tenant controls API",
    "version": "2026-09-09.2",
    "description": "Phase 11 tenant-operation supplement. Existing business operations remain documented in the application API reference and phase contracts. This is not yet the final whole-product parity inventory."
  },
  "paths": {
    "/v1/tenant/clients": {
      "get": {
        "operationId": "listCommunicationsClients",
        "description": "50 managed single-tenant clients per key-cursor page. Legacy and multi-tenant operator credentials are excluded.",
        "security": [
          {
            "ApiKey": [],
            "Tenant": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative app-owned result; not proof of an external provider effect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Scope, role or membership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or held operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "changeCommunicationsClient",
        "description": "",
        "security": [
          {
            "ApiKey": [],
            "Tenant": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative app-owned result; not proof of an external provider effect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Scope, role or membership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or held operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "operation",
                  "keyId"
                ],
                "properties": {
                  "operation": {
                    "enum": [
                      "create",
                      "rotate",
                      "revoke"
                    ]
                  },
                  "keyId": {
                    "type": "string",
                    "pattern": "^[a-zA-Z0-9_-]{8,80}$"
                  },
                  "revision": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "secret": {
                    "type": "string",
                    "writeOnly": true,
                    "minLength": 32,
                    "maxLength": 256
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "enum": [
                        "admin",
                        "reader",
                        "writer"
                      ]
                    }
                  },
                  "capabilities": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "description": "Administrator plus tenant:manage and communications read/write capabilities required. Creation delegates only supported caller capabilities to one tenant. Rotate/revoke require current client revision. Identical create/rotate replay is safe; revoked clients cannot be reactivated. Caller supplies and retains the secret; responses omit secrets and hashes."
              }
            }
          }
        }
      }
    },
    "/v1/tenant/usage": {
      "get": {
        "operationId": "readCommunicationsUsage",
        "description": "90-day view of managed authenticated business API requests. Excludes legacy credentials, the three tenant administration endpoints and provider charges.",
        "security": [
          {
            "ApiKey": [],
            "Tenant": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative app-owned result; not proof of an external provider effect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Scope, role or membership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or held operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "setCommunicationsRequestBudget",
        "description": "Administrator plus tenant:manage. Limit zero means uncapped. Atomic UTC daily request admission; administration remains accessible for recovery.",
        "security": [
          {
            "ApiKey": [],
            "Tenant": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative app-owned result; not proof of an external provider effect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Scope, role or membership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or held operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "revision",
                  "dailyLimit"
                ],
                "properties": {
                  "revision": {
                    "type": "integer"
                  },
                  "dailyLimit": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/audit": {
      "get": {
        "operationId": "readCommunicationsTenantAudit",
        "description": "Administrator plus tenant:manage. 50 scoped audit records ordered by time and ID. Records creation, rotation, revocation and budget changes atomically; a live offset view, not a frozen export.",
        "security": [
          {
            "ApiKey": [],
            "Tenant": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative app-owned result; not proof of an external provider effect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Scope, role or membership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or held operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000
            }
          }
        ]
      }
    },
    "/v1/tenant/lifecycle": {
      "get": {
        "operationId": "readCommunicationsLifecycle",
        "description": "Read local lifecycle state and the export catalogue; dataset requests require memory:private and a current suspended revision. Credential stores are excluded, capability URLs and authentication fields are redacted. At most 50 records and 3.5 MB per page. This is not an operator backup.",
        "security": [
          {
            "ApiKey": [],
            "Tenant": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative app-owned result; not proof of an external provider effect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Scope, role or membership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or held operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Export page exceeds the bounded REST response; operator file export required"
          }
        },
        "parameters": [
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "post": {
        "operationId": "changeCommunicationsLifecycle",
        "description": "Administrator tenant:manage only. Stable request IDs replay the original receipt. Suspension rejects held provider operations. erase_local additionally requires tenant:erase and the exact confirmation; it retains a tombstone and audit/usage, and does not erase HyperFlow or provider resources.",
        "security": [
          {
            "ApiKey": [],
            "Tenant": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authoritative app-owned result; not proof of an external provider effect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Scope, role or membership denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or held operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request budget exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "operation",
                  "revision",
                  "requestId"
                ],
                "properties": {
                  "operation": {
                    "enum": [
                      "suspend",
                      "resume",
                      "erase_local"
                    ]
                  },
                  "revision": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "requestId": {
                    "type": "string",
                    "pattern": "^[a-zA-Z0-9_-]{8,100}$"
                  },
                  "confirmation": {
                    "type": "string",
                    "const": "Erase Communications local tenant data"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "Tenant": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-Id"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}